Ignore media source intent on config change
Do not process the original intent to set the media source, when
activity is re-created in responce to the configuration change. This
makes sure the media source is not swapped back to the original one,
when configuration change happens, if the user has switched to a new
source before it.

To check the fix, choose a random media source from the AppGrid, switch
to another one using the source selector and then trigger a
configuration change (e.g., apply an RRO to change some colors, which
also targets the Media Center). The source should retain.

Bug: 294088465
Test: manual (see steps above)
Change-Id: I4122c75e6e40ebf55dbb12ac0b27c0b097038c7c