Fix MediaActivity leaks in CarMedia app
Forcibly remove an infinite observer holding an activity reference and linked
to a `ViewModel`, which causes a `MediaActivity` activity leak, when the
activity is destroyed.

Use application context for the retained `MEDIA_SOURCE_MODE_BROWSE`
`MediaModels` instances, since the context reference is held by the
object and causes the leak of the first `MediaActivity` instance, which
is used for initial set-up.

Bug: 317324773
Test: manual - triger a few unhandled configuration changes and check memory
profiler for detected leaks

Change-Id: Ieb89b746ff50c5588365e2accebd1e02a06ae1c6